+++ types/kafkajs.d.ts
@@ -430,6 +430,12 @@ export type Admin = {
       timeout?: number,
       isolationLevel: IsolationLevel
     }): Promise<Array<SeekEntry & { high: string; low: string }>>
+  fetchTopicOffsetsByTimestamp(topic: string,
+    timestamp?: number,
+    options?: {
+      timeout?: number,
+      isolationLevel: IsolationLevel
+    }): Promise<Array<SeekEntry>>
 }