+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# confluent-kafka-javascript v1.1.0
+
+v1.1.0 is a feature release. It is supported for all usage.
+
+## Enhancements
+
+1. Add support for an Admin API to fetch topic offsets by timestamp (#206).
+
+
 # confluent-kafka-javascript v1.0.0
 
 v1.0.0 is a feature release. It is supported for all usage.